requirements.txt
telethon
pysocks